% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_test}
\alias{d_test}
\title{Simple dictionary for the test dataset.}
\format{
A List with a single entry:
}
\usage{
d_test
}
\description{
A List containing mapping of HLA ambiguity codes to all possible alleles.
}
\keyword{datasets}
