% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Extensive dictionary for the example dataset.}
\format{
A List with a 17930 entries:
}
\usage{
dict
}
\description{
A List containing mapping of HLA ambiguity codes to all possible alleles.
Some ambiguities are situational and do not correspond to official MAC;
those are denoted by UNK#### with 4 digits numeric identifiers.
}
\keyword{datasets}
