% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egf}
\alias{egf}
\title{List of data.frames of allele frequencies for the example dataset.}
\format{
A List with 4 loci ("A", "B", "C", "DRB1").
}
\usage{
egf
}
\description{
A List with each locus as a list element. For each locus a data.frame with
the alleles as rows and different regions or populations as columns.
}
\keyword{datasets}
