% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egh}
\alias{egh}
\title{Haplotypes frequency for the example dataset.}
\format{
A data frame with 4381 rows and 3 variables:
\describe{
  \item{alleles}{the haplotypes with alleles separated by "~"}
  \item{NA.}{frequency for the example data}
  \item{NA.smpl}{frequency for the simplified example data}
  ...
}
}
\usage{
egh
}
\description{
A dataset with two different populations (full and simplified) according to six loci haplotypes.
}
\keyword{datasets}
