% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h_test}
\alias{h_test}
\title{Haplotypes frequency for the test dataset.}
\format{
A data frame with 7 rows and 3 variables:
\describe{
  \item{alleles}{the haplotypes with alleles separated by "~"}
  \item{01}{frequency for the test data}
  \item{02}{frequency for the simplified test data}
  ...
}
}
\usage{
h_test
}
\description{
A dataset with two different populations (full and simplified) according to four loci haplotypes.
}
\keyword{datasets}
