% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t_test}
\alias{t_test}
\title{Sample genotypes for the test dataset.}
\format{
A data frame with 4 rows and 10 variables:
\describe{
  \item{IND}{the individual identifier}
  \item{A_1}{first HLA-A allele}
  \item{A_2}{second HLA-A allele}
  \item{B_1}{first HLA-B allele}
  \item{B_2}{second HLA-B allele}
  \item{C_1}{first HLA-C allele}
  \item{C_2}{second HLA-C allele}
  \item{DRB1_1}{first HLA-DRB1 allele}
  \item{DRB1_2}{second HLA-DRB1 allele}
  \item{CRA}{region/population of the individual}
  ...
}
}
\usage{
t_test
}
\description{
A dataset with sample genotypes (4 loci).
}
\keyword{datasets}
